Don't lock pipelines during local layout update
In our Zuul deployment we can see that the local layout update takes
very long on a busy scheduler. We also observed that due to the locking
in the post-config hook, we might starve pipelines for the duration of
the local layout updates.

Since the pipeline state is already reset after the inital
reconfiguration we don't need to reset it again after a local layout
update. In case the pipeline's layout UUID is already the same as our
current one, we can skip that step.

This change also removes the relative priority queues from the pipeline
state. The relative prio queues were never properly (de-)serialized and
shouldn't be stored in ZK at all.

Change-Id: I7ad003a8fc24ab9be417a46287390675227e2db3