Clone from /a/ with authenticated Gerrit HTTP
When using authenticated HTTP with Gerrit (instead of SSH) for
cloning repos, use the "/a/" URL prefix.

Gerrit supports cloning with or without the prefix going back to
at least 2.16.  However, if users install the login-redirect plugin,
it disallows HTTP access except via "/a/".

Further, the download-commands actually do recommend using /a/ in
clone URLs.

Switching to this should help users with login-redirect installed
without any loss of functionality otherwise.

Change-Id: I9b65d0b9a0f235afe076a34a9e878f1e8fefe1a0