Always send dequeue event to superseded pipelines
The pipeline manager was accessing other pipelines without locking and
refreshing them when trying to dequeue superseded items. This could have
lead to data races if the pipeline state was outdated.

There is still a race condition here if the pipeline with the supersedes
is processed first:

0. trigger event created for 'check' and 'gate' (supersedes 'check')
1. enqueue change in 'gate'
2. dispatch dequeue event for 'check'
3. process dequeue event for 'check' -> exception as change not in
   pipeline
4. enqueue change in 'check'

A possible fix for this might be to also add the inverse relationship to
the superseded pipeline in order to check the other pipelines change
list before enqueueing a change.

Change-Id: I5cee72ebc82e5078b794449eb0e129dba65ff2c4