Don't add changes already in superceding pipeline
This change fixes a race condition related to trigger event processing
between a pipeline that supersedes another one.

The race condition can happen as follows:

0. trigger event created for 'check' and 'gate' (supersedes 'check')
1. enqueue change in 'gate'
2. dispatch dequeue event for 'check'
3. process dequeue event for 'check' -> exception as change not in
   pipeline
4. enqueue change in 'check'

In order to fix this we will check if the change is already in a
pipeline that supersedes the current one. For this we need to store the
inverse relationship for the `supercedes` attribute as
`superseded_by` during config loading.

When adding a change we'll first check the if the change is already in
any of the pipelines it is superseded by. For that the pipelines change
list will be used.

Change-Id: I62627608e55202e70761ef07ffeb9a52eb09ef6e