Fix leaked semaphore cleanup
The cleanup method of the semaphore handler was accessing the pipelines
without locking and refreshing them. This could lead to the cleanup of
semaphores that were still valid.

To fix this we will change the semaphore handle from
`<item_uuid>-<job_name>` to a dictionary:

    {
        "buildset_path": <buildset-path>,
        "job_name": <job_name>,
    }

In the cleanup routine we can then check if the buildset path still
exists. If that's not the case we know that the semaphore was leaked and
can clean it up.

To be backward compatible we will still allow the legacy format of the
handle when acquiring and releasing a semaphore. In the cleanup method
we will ignore the legacy semaphore handles and just log a warning
message.

This change also updates the model API version to 2. To be backward
compatible we will still use the old-style handle format as long as
there are still components with model API <2. We'll also not execute the
cleanup routine in that case.

Change-Id: I8b0706818b0c2793aae11b8d85cc8ff2d70b28fc