Ignore KeyErrors in node cache listener deleted events
Currently, the node cache listener raises a KeyError when it receives a
NODE_REMOVED event for a node that is not in the cache dict [1].

The only action in the node cache listener for a NODE_REMOVED event is
to remove the node from the cache. As it is already gone, we could
simply ignore this exception.

[1]:
2022-02-06 04:34:14,838 ERROR zuul.zk.nodepool.ZooKeeperNodepool: Exception in node cache update for event: (2, ('/nodepool/nodes/0039931194', b'', ZnodeStat(czxid=345614355535, mzxid=345614355535, ctime=1644122054794, mtime=1644122054794, version=0, cversion=1, aversion=0, ephemeralOwner=0, dataLength=0, numChildren=1, pzxid=345614355538)))
Traceback (most recent call last):
  File "/opt/zuul/lib/python3.8/site-packages/zuul/zk/nodepool.py", line 572, in nodeCacheListener
    self._nodeCacheListener(event)
  File "/opt/zuul/lib/python3.8/site-packages/zuul/zk/nodepool.py", line 620, in _nodeCacheListener
    node = self._node_cache.pop(node_id)
KeyError: '0039931194'

Change-Id: Idedaca153b82b288f112902293da9ae5ca834422