Avoid data races when dequeueing superceded items
The pipeline manager was accessing other pipelines without locking and
refreshing them when trying to dequeue superceded items. This could have
lead to data races if the pipeline state was outdated.

To fix this, we will send a new `SupercedeEvent` to all pipelines that
are configured as supercedes. This event will go into the pipeline's
*trigger* event queue instead of the management event queue (as was the
case for the `DequeueEvent`, that we used before). This way we can avoid
race conditions with still in-flight trigger events whose associated
change would otherwise not be dequeued.

This updates the model API version to 2 due to the new event.

Change-Id: Id390ddd95b27c3fabda14c07a524a1c6f71ef72f