Update patch set 1
Patch Set 1:

It looks like this may be able to happen if the treecache gets a created and deleted event rapidly in succession.  I think we should remove our use of TreeCache and align this with the way we do our own caching in the rest of the system.  We're not really using the features of TreeCache.

Patch-set: 1
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Attention: {"person_ident":"Gerrit User 27952 \u003c27952@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}