Make a global component registry
We generally try to avoid global variables, but in this case, it
may be helpful to set the component registry as a global variable.

We need the component registry to determine the ZK data model API
version.  It's relatively straightforward to pass it through the
zkcontext for zkobjects, but we also may need it in other places
where we might alter processing of data we previously got from zk
(eg, the semaphore cleanup).  Or we might need it in serialize or
deserialize methods of non-zkobjects (for example, ChangeKey).

To account for all potential future uses, instantiate a global
singleton object which holds a registry.  The first component
registry constructed will be automatically attached to it and it
will be available anywhere.  We also add a clear method so that
we can be sure unit tests start with clean data.

Change-Id: Ib764dbc3a3fe39ad6d70d4807b8035777d727d93