Update patch set 9
Patch Set 9: Code-Review-1

(2 comments)

The test failure shows that this does have an incidental dependency on the semaphore change.  By optimizing this case out, we can end up with empty queues on scheduler start, at least until the first pipeline process is run.

That shows up in the semaphore test since it's relying on out-of-date information.  The semaphore fix fixes that bug, so if we stack this on top of that, the test works.  But the failure is instructive in that it's altering some assumptions.  I think we should consider how to proceed.

More detailed comments inline.  But to summarize my recommendation:

* Accept the behavior change
* Remove the initial current_changes stats call
* Merge this after the semaphore fix

Patch-set: 9
Label: Code-Review=-1
Attention: {"person_ident":"Gerrit User 1 \u003c1@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"REMOVE","reason":"removed on reply"}