Better document service command-line switches
Improve the command-line option info at the start of the Operation
page to be a bit more verbose about -f in order to make it easier to
keyword search for, and include information on -d as well. Also link
to this from the Troubleshooting page, since increasing log detail
is likely to be one of the first steps before using the lower-level
tools mentioned there.

Change-Id: I369215df64a2dae22baa65cd010baaa3ec0b2c30