Update patch set 3
Patch Set 3:

It looks like the quickstart failures may be related to slow zuul-web startup, though all the new loglines added by this change are at debug level so it's hard to tell where in the start() method things are taking so long.

Should we increase log verbosity? Is only this change impacted?

Patch-set: 3