Fix a bug in getting changed files
The original implementation takes into account the changed fils from
all commits of a PR.

It causes a bug when files get changed and reverted in those commits.
e.g. A file is added in first commit then removed in second commit,
this file should should not be considered as a changed file in the PR.

The fix will use diff() files to filter out those files that are not
actually changed.

The reason we do not use the result of diff() is, it can possibly include
files that are not part of the change  when the reference branch has diverged
from the feature branch.

Change-Id: I15a064983c74c4b4bc90ab312953a0e7170c1326