Fix a bug in getting changed files
The original implementation takes into account the changed fils from
all commits of a PR.

It causes a bug when files get changed and reverted in those commits.
e.g. A file is added in first commit then removed in second commit,
this file should should not be considered as a changed file in the PR.

The fix tries to find the commen base first, and then find the diff
files between the feature branch and the common base.

Change-Id: I15a064983c74c4b4bc90ab312953a0e7170c1326