wip: release tenant read lock early in case of reconfig
With large tenants and during high load it can take quite some time
until all schedulers have released a tenant read lock acquired during
pipeline processing.

In order to release the tenant read lock earlier we will check the lock
contenders for a reconfig identifier before processing a pipeline.

If a reconfig is waiting to acquire the write lock we will release the
read lock earlier and continue with the next tenant.

Change-Id: Ibfde3c23104e1f0ea788326d250619dfcd29e68c