Release tenant read lock early on pending reconfig
With large tenants during high load phases it can take quite some time
until all schedulers have released a tenant's read lock that is acquired
during pipeline processing. This means that a scheduler that is trying
to reconfigure a tenant might need to wait several minutes in order to
acquire the tenant write lock. During this time the run handler of that
scheduler is completely blocked. The pending write lock also prevents
other schedulers from starting to process that tenant in the meantime.

Because of that we want to release a tenant's read lock as soon as
possible so the pending reconfig can proceed.

In order to release the read lock earlier we will check the lock
contenders for a reconfig identifier before starting to process a
pipeline. This identifier needs to be passend when a scheduler tries to
acquire a tenant's write lock.

Change-Id: Ibfde3c23104e1f0ea788326d250619dfcd29e68c