Add buildset start/end db columns
Add two columns to the buildset table in the database: the timestamp
of the start of the first build and the end of the last build.  These
are calculated from the builds in the webui buildset page, but they
are not available in the buildset listing without performing
a table join on the server side.

To keep the buildset query simple and fast, this adds the columns to
the buildset table (which is a minor data duplication).

Return the new values in the rest api.

Change-Id: Ie162e414ed5cf09e9dc8f5c92e07b80934592fdf