Report gross/total tenant resource usage stats
Export a new statsd gauge with the total resources of a tenant.
Currently, we only export resources of in-use nodes. With this, we
additionally report the cummulative resources of all of a tenants nodes
(i.e. ready, deleting, ...).

This also renames the existing in-use resource stat to distinguish those
clearly.

Change-Id: I76a8c1212c7e9b476782403d52e4e22c030d1371