Don't submit empty node requests to Zookeeper
We are currently also submitting empty node requests to Zookeeper. As
far as I could see that was only necessary as an intermediate step
towards the scale-out scheduler. Since we have all state in Zookeeper by
now it seems that we don't need this step anymore.

By that we can also save one run handler cycle for jobs without any
nodes. This is done by setting the nodeset on the buildset in case the
request is already fulfilled immediately after requesting nodes.

Change-Id: I4f9cbc7555591bb8817e3596edf4b9af99efd998