Don't submit empty node requests to Zookeeper
We are currently also submitting empty node requests to Zookeeper. As
far as I could see that was only necessary as an intermediate step
towards the scale-out scheduler. Since we have all state in Zookeeper by
now it seems that we don't need this anymore.

Since we will no longer receive a nodes provisioned event for the empty
node request, we will set the empty nodeset immediately after requesting
the nodes if the request is fulfilled at this point.

By not submitting empty node requests to Zookeeper, we can also save one
run handler cycle for jobs that don't need any nodes.

Change-Id: I4f9cbc7555591bb8817e3596edf4b9af99efd998