Protect stream handler thread from exceptions
Change I6da1c88dfe20eb9e1ada09d7aa741f9024ddfc04 updated the log
streaming handler to explicitly close streaming sockets rather than
waiting for them to be garbage collected.  However, it appears it
may be possible now for the unregisterStreamer method to be called
after the socket is closed.  Perhaps it can happen if the socket is
closed due to an error on transmission.  The exact mechanism is not
clear, but the following errors were observed:

:Exception in thread StreamManager:
Traceback (most recent call last):
  File "/usr/local/lib/python3.8/site-packages/zuul/web/__init__.py", line 1647, in run
    streamer.handle(event)
  File "/usr/local/lib/python3.8/site-packages/zuul/web/__init__.py", line 340, in handle
    self.websocket.send(data, False)
  File "/usr/local/lib/python3.8/site-packages/ws4py/websocket.py", line 303, in send
    self._write(m)
  File "/usr/local/lib/python3.8/site-packages/ws4py/websocket.py", line 285, in _write
    self.sock.sendall(b)
BrokenPipeError: [Errno 32] Broken pipe

During handling of the above exception, another exception occurred:

Traceback (most recent call last):
  File "/usr/local/lib/python3.8/threading.py", line 932, in _bootstrap_inner
    self.run()
  File "/usr/local/lib/python3.8/threading.py", line 870, in run
    self._target(*self._args, **self._kwargs)
  File "/usr/local/lib/python3.8/site-packages/zuul/web/__init__.py", line 1651, in run
    self.unregisterStreamer(streamer)
  File "/usr/local/lib/python3.8/site-packages/zuul/web/__init__.py", line 1675, in unregisterStreamer
    self.poll.unregister(streamer.finger_socket)
ValueError: file descriptor cannot be a negative integer (-1)

While the exact sequence is not clear, the following should both
prevent the issues from recurring and help produce accurate logs
in the future:

1) Preserve the fileno on the stream handler so that we can unregister
it from the poll even after it has been closed.  We catch KeyErrors
from poll.unregister, so if we attemp to unregister it twice, that's
fine.  We just can't unregister -1.  So if we save the fileno that we
use when registering and use the same one when unregistering, we should
avoid any attempts to unregister an invalide fileno.

2) Put the entire stream handler run loop in a try/except.  This is
a pattern we generally use to avoid errors like this killing threads
which otherwise should continue running.

Change-Id: I67e67c0e1406cba2d31af28af0dcba302003af81