merger: ignore a missing object when creating refs
We recently hit a problem where an upstream GitHub project rebased a
backport change, then deleted the PR branch, leaving us trying to set
refs/heads/backport/blah/blah to an object that had now disappeared.

Unfortunately this was impossible for the submitter to diagnose as the
only indication of this was an exception from
merger/merger.pY:_setRef() at

 obj = git.objects.Object.new_from_sha(repo, binsha)

which was swallowed in the executor logs, and resulted in the job just
failing with RETRY and no logs.

My feeling is that we can ignore this missing object at this point and
log.  Later on, if we are actually trying to use this branch (via a
depends-on) we will get a better error message if it doesn't exist in
the on-disk repos.

Change-Id: Id395b1e0d2c445ae89b0921436d59c36ae4cb155