merger: report missing object as ERROR
We recently hit a problem where an upstream GitHub project rebased a
backport change, then deleted the PR branch, leaving us trying to set
refs/heads/backport/blah/blah to an object that had now disappeared.

Unfortunately this was impossible for the submitter to diagnose as the
only indication of this was an exception in the executor logs.

Catch this error and report it back.

Change-Id: Id395b1e0d2c445ae89b0921436d59c36ae4cb155