Fix error in background tenant update of new tenant
When adding a new tenant, the background update thread can throw
a KeyError because we don't have a local state for the new tenant.
Protect from that and add a test of this sequence.

Do the same in the main run loop as well.  Also, add a tenant-level
catch-all exception handler so that if there is a problem with a
tenant, we continue processing other tenants.

Change-Id: I9720e903c776fa009e8bc77045e42a0b3ab1d326