Don't reconfigure after every gitlab merge
The gitlab driver sets change.files to None on branch push events.
Like every other system, it also emits branch push events after
merging merge requests.  When the scheduler encounters a branch
updated event with files.None, it assumes the worst and performs
a tenant reconfiguration.  This means that Zuul performs a tenant
reconfiguration after every single MR is merged.

Gitlab supplies a list of changed files on push events.  This change
modifies the gitlab driver to do that.  This mirrors what is done
in the github driver.

Change-Id: Ia6651048eea7f183ee0a65b6e8de83fdf892c9a7