Add "repo state" waiting status
The "waiting" status in the web ui represents the time between
when the job graph is frozen (and therefore we can show the job
list) and the time the node request is enqueued.  If we're waiting
on other jobs, we show a "waiting status" mouseover, but if we're
waiting on the global repo state merge job, we display nothing.

Since that can take some time, add that as a waiting status.
Now the only time we will display "waiting" without a mouseover
status should only be for very brief times of internal processing
within a single scheduler run loop.

Change-Id: I85f2b08da6350a5f7339fc9072e5fe1b01b51a84