Add waiting status to queued
We currently supply no "waiting status" information (which would appear
in the mouseover text on the status page) for jobs that are queued.
This represents jobs for which we have submitted a node request but the
build has not started.

This change adds two new "waiting_status" values: the node request id
for the period between node request submission and executor build
request submission, and the string "executor" for the period between
executor build request submission and build start.

This updates the status page to display waiting_status as mouseover
text for the "queued" status label.

Change-Id: Ic1151944f4a8da8d1ab0a3ae96efb0877f86aebd