Report which repo failed initial merge ops
When the initial merge job for a queue item fails, users typically
see a message saying "this project or one of dependencies failed
to merge".  To help users and/or administrators more quickly identify
the problem, include connection project and change information in
a warning message posted to the code review system.

Change-Id: If1bced80b87b908f63867083efb306ebe02ed1ee