Fix removal of tenant on layout update
Removed tenants where not correctly handled by the layout update logic
in the scheduler and zuul-web. The tenant was only removed from the
scheduler that process the reconfiguration event but not from other
scheduler and web instances that reacted on the layout update.

In addition to that we now also clear the TPCs of the removed tenant.

Change-Id: I0b3ff77388c27132a558207316bd42c8646212c6