Add max retries to refresh of pipeline change list
This adds a limit (5 retries) on the number of times we retry to refresh
a pipeline's change list. In case of a corrupt change list a scheduler
might hang in the refresh for any length of time.

We also observed an issue where a scheduler was unblocked in such a case
after a few hours and cleaned up the change cache based on outdated data
that was collected before the hanging refresh.

Change-Id: I866fefde8252da00e07c418b17957451dc914540