Only process pipelines if there are outstanding events
While the overall system throughput is higher with multiple schedulers,
we have not yet optimized thair load-sharing, so in some cases the system
may not be operating as efficiently as it could.

Consider the sequence of processing tenants and pipelines.  Currently
once the main loop wakes up (due to an event being added) it iterates
through all the tenants and within each tenant, all the pipelines and
processes their queues.  If any pipeline is locked by another scheduler,
it skips it.

This can be very efficient if the schedulers manage to alternate:

1: process pipeline A
2: skip pipeline A, process pipeline B
1: skip pipeline B, process pipeline C
2: skip pipeline C
etc.

But can be very inefficient if one simply follows the other:

1: process pipeline A
2: process pipeline A
1: process pipeline B
2: process pipeline B

In that case, it should be possible for scheduler 2 to determine that it
does not need to process a pipeline which was just processed.  Nothing*
should change within a pipeline without a corresponding event,
so before refreshing a pipeline from ZK and processing it, we can check
the pipeline event queues and ensure that there is work to do.  If
scheduler 1 has already processed it, the queues should be empty and we
can skip it.  Otherwise we should process it.

* The exception to this is that a reconfiguration can adjust the
configuration of a pipeline which will then need all its items
re-enqueued.  We check whether a pipeline state has old_queues during
processing for this.  We still need that to run after a reconfiguration
and there may not be a pipeline-specific event to trigger the run with
the proposed scheme.  To address this, a new PipelinePostConfig event
is added merely so that there will be an event that triggers a pipeline
processing run in this situation.

Both the processing and bypass code paths are exercised extensively by
the existing tests.

The setOldQueues method is unused and is removed.

This change uncovered a bug related to a missing timestamp in
ZuulTriggerEvent.  This caused an exception in the SQL reporter which then
necessitated an extra pipeline process to correct.  This caused the
test_zuul_trigger_parent_change_enqueued tests to fail.  This has been
corrected by adding the timestamp.

Change-Id: Ie3bd3b1761f676868b766ffb915cf9c3a8c99618