Items in extra paths should be loaded in dependent changes
If a change depends on a change of another repo, who has
extra-dirs/files, the items were not loaded from those extra paths,
because only the extra_config_files/dirs of the current repo was
considered.

Fix the issue by also take into account all the extra_config_files
/dirs from the all the dependent repos.

A side effect would be, it would try to load items in all the extra
paths in all the repos. Potentally unwanted paths could be looked up
for items.

Change-Id: Ia50eed6c2c11b392fe83469453546a4b0928a769