Add a "dirty" flag to pipeline processing
If an error causes us to abort pipeline processing before the queue
is stabilized, but after we have consumed all of the events, we might
sit idle until a new event arrives.  To avoid that and maintain the
previous behavior of aggressively running the pipeline processor as
until it completes without error, set a "dirty" flag when we start
and clear it when we finish.  The presence of the dirty flag indicates
that we should run the pipeline processor regardless of whether there
are any events.

Change-Id: I3581bc9ad08b208b7658e152342a651c2cd54ce4