Use worker zone for log streaming from component
Currently, we are looking up the worker_zone for the log streaming from
the BuildRequest's path in ZooKeeper. This is a problem for unzoned
builds as those builds don't provide a zone information in their path
(zone=None).

Due to this, the log streaming won't use the FingerGateways and instead
always falls back to use the direct connection to the executor. This
works as long as the executor is located in the same region as zuul-web,
but in other cases the log streaming is broken.

To fix this, we will look up the worker zone from the component registry
and use that instead of the zone extracted from the build request path.

Co-authored-by: Felix Edel <felix.edel@bmw.de>
Change-Id: Ia5ed9109866131ae8549908bac687287e7625c61