Fix multi-scheduler test races in waitUntilSettled
Several of the calls in waitUntilSettled accept a scheduler matcher
but probably shouldn't.  For example, the check that all node reqs
have been handled assumes that it's examining the scheduler that
both received the node request completed notification and put the
resulting event on the queue.  If it was a different scheduler, then
the two data structures may be out of sync.

The matchers are useful for only checking event queues and run handler
locks for a single or subset of schedulers.  Reduce the method to
only include matchers for those cases and check all the schedulers
for the items where it's necessary.

Change-Id: I087bc8758011388788544b865262a6d345e71472