Use kazoo.retry in zkobject
It appears that the division of retryable/non-retryable exceptions
is not exactly as asserted in the comments in the retry handlers
in zkobject.  The kazoo.retry module has a class that implements
retry handling with the correct set of retryable exceptions, so
let's use that instead.

The main thing we lose here is the log messages indicating we are
in a retry loop.  There doesn't appear to be a good way to hook
into KazooRetry for that (we could log in the interrupt method,
but that is called every 0.1 seconds).

The InvalidObjectError exception appears to be unused, so it is
removed (since continuing to use it would make the exception handling
more complex).

Change-Id: I1278cd27873374b4efd90504d7166c74c6057b52