Correctly reset failing cycle behind failing item
In case a cycle and another item ahead failed in the same pipeline run,
the circular dependency wasn't correctly reset in some cases.

Instead of resetting the items in the cycle and moving it behind the
nearest non-failing item (nnfi), it was moved behind the failing item.
Only when the failing item reached the head of the queue it lead to a
gate reset which corrected the pipeline state. Since the gate reset was
delay this can lead to wasted build resources.

To fix this we will only consider failing dependent (bundle) items ahead
in the queue and ignore items behind.

Change-Id: I3f11a917028b2ad58403a81d6542c6166bb3a663