Refresh pipeline change list and status when refreshing the state
The ZKObject uses an internal hash to determine whether it has been
updated since the last time it was read so that we can avoid unecessary
ZK writes.  Therefore we must ensure that we read the latest data
before we make any updates.  However, we were not doing that for the
pipeline summary.  That could permit the following sequence:

* Scheduler A initializes the pipline summary (empty)
* Scheduler B enqueues an item in the pipeline and updates the summary
  to include the item
* Scheduler A removes the item and updates the summary to exclude
  the item; but that is the same content as it had previously so
  the ZK write is skipped

We are left with a summary with the item included in the pipeline
when it actually is not.

Note that the attached test runs no jobs -- that's because in order
for this case to hit, the change must be added and removed from the
pipeline in only two cycles, which is only possible with projects
that do not participate in the pipeline.

To correct this, refresh the pipeline summary when we refresh the
pipeline state.  Additionally, do the same for the change list; though
the problem is much less likely to occur in this case since the
change list is refreshed during tenant trigger event forwarding which
occurs immediately prior to pipeline processing.  But there is still
a window where we could refresh the change list, and another scheduler
could process a pipeline in the interim.  This is more defensive.

Change-Id: I365c19ccb567a53f35c68bfd03acf42e8739345a