Calculate elapsed and remaining times in javascript
The scheduler may now skip pipeline processing if there are no
pending events.  And we only write the status JSON to ZK when we
process pipelines.  That means that the JSON received by the web
app may be very old.  Everything in it should still be correct
except for the calculated values such as elapsed and remaining time.

This means that if a tenant is less active, the times and progress
bars may appear frozen.  To mitigate this, use the absolute timestamps
and calculate the relative times on the client.

Change-Id: Ia2bf8bb3b5495e3db13723f81c44186e5ca6ff53