Check if job request is still queued before yielding
When iterating over job requests, we can end up yielding requests
which are no longer present because they have been completed, or
are in progress by another merger or executor.  We can't catch
all of those cases, which is why we try to lock it and if the lock
fails, move on to the next.  But we can catch some of them and
avoid the attempted lock (and therefore extra network traffic and
warning messages).  This change double checks the current state
inside the generator before yielding.

Change-Id: I7e2b319a79f6a608ffaf261e65e9532e3486a47f